fix(be): fix check team null trong TeamDAO

+++ DataAccess/TeamDAO.cs
@@ -79,6 +79,7 @@ namespace DataAccess
         }
         public async Task<bool> UpdateAsync(Team team)
         {
+            if (team == null) throw new ArgumentNullException(nameof(team));
             _context.Teams.Update(team);
             return await _context.SaveChangesAsync() > 0;
         }
@@ -108,6 +109,8 @@ namespace DataAccess
 
         public async Task DeleteAsync(Team team)
         {
+            if (team == null) throw new ArgumentNullException(nameof(team));
+
             if (team.Teammembers.Any())
                 _context.Teammembers.RemoveRange(team.Teammembers);
